feat(ui): TE-1270 moved subscription view card to entity-cards folder
