+++ thirdeye-ui/src/app/components/entity-cards/subscription-group-view-card/subscription-group-view-card.interface.ts
